requirements.txt
streamlit==1.41.1
langgraph==0.6.4
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.27
langchain-openai==0.3.0
ollama==0.5.3
tiktoken==0.8.0
python-dotenv==1.0.1
openai